rendfont.cpp,rendlay.cpp: add font support to `<text>` components
This adds a "font" attribute to the `<text>` layout component, defaulting to `"default"` if nothing is specified.

It also changes the `render_font` constructor to accept not just a single "filename" but a comma-separated list of "filenames". These actually don't seem to be file names at all, but font specifications that may be symbolic (like the default `"Liberation Sans|Regular"` in `font_osd.cpp`), which includes a `|` separator between font family and style.

This is in turn used in the Ensoniq VFX family of keyboards to use the ccorrect Bold and Italic styles of a suitable Helvetica-compatible sans-serif font, falling back to "default" of nothing is found. This includes a first choice of an explicitly nonexistent font to demonstrate that the list of fonts is in fact traversed.

This also includes support for generic font families.